Restructure BAHAMAS read in
Since all time steps of the flight are needed, it makes more sense to read in the whole BAHAMAS data set for the flight, making the get_position function useless. It is kept for potential later use.

The new read_bahamas function takes the flight as an argument and builds the path itself. However, the full path to the ncfile can still be provided and overwrites the standard path.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>