Terminate infinite while loop
If the libRadtran simulation does not compute because of an input error (i.e. negative values in the radiosonde file) the file_check loop will run indefinitely. This will put it in a try except clause and after ten iterations will automatically throw an error which terminates the while loop and jumps to the next iteration of the for loop.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>