Add logging to file
Instead of piping the StreamHandler output to stdout and then into a log file via tee in the command line, this will create a FileHandler from the script name and pipe all output to stdout there.
In that way the progress bar from tqdm which is sent to stderr does not globber up the log file.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>