Setup new project structure
In order to make those scripts and functions useful for others a new module pylim is created and the scripts are sorted by their respective tasks. The idea is to have functions in the module and then processing scripts which are in charge of calibrating data and writing them to netCDF files. The quicklook scripts are for online quicklook generation during the campaign. analysis is meant to be a folder for all things which generate knowledge from the data. Each script starts with an instrument name to signal which instrument data is being used. Multiple instruments can be used as well. Then the main instrument of interest defines the file name. For general campaign analysis the respective campaign name is used in the script.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>