Use absolute path for config file
When executing get_path config.toml was expected to be in the current directory. With the new module structure this does not work anymore when get_path is called during the execution of another script. It still works in the console though.

Using the pathlib module get_path now finds the project directory, which should be the parent directory of the pylim module, and creates an absolute path to config.toml which should be in the project directory.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>