Implement new dark current correction for SWIR
Try to use the diffuse measurement to the corresponding angle and take the mean of that measurement to correct the dark current of the current file. If now diffuse measurement is available, take the closest available folder and search for a file there. If that does not work loop through all diffuse folders until a file is found. -> To be replaced by the 0° measurement

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>