Move pylim to src folder
To be in line with the recommended folder structure for python modules pylim was moved to src. This required pylim to be properly installed by pip in order to be able to import it via import pylim.
That also messes up the path to the config.toml file which is checked for in the init file. Thus, this check is removed and the get_path() function now works on the current working directory as a root folder for the config file.

In addition, the import of pylim inside the module is moved to the top of each import block.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>