Update pylim to version 0.2.1
Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>

Add new version to dark current calculation

In get_dark_current() a new way to calculate the dark current for the VNIR channel is introduced. Using the 19th to 99th pixel from the spectrometer the measured dark current is scaled to the actual measurement.

Additionally, the optional plot of the dark current is called at the end of the function. Therefore, the provided x-axis (the dark current wavelengths) are saved to the variable wsl in each if section.

Furthermore, simple logging info is changed to Errors which are raised to stop the program instead of letting it run.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>

Update requirements.txt

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>

Add ozone sonde functions and dictionaries

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>