Add variable wavelength
In order to run different settings of libRadtran with less need to change things a variable called wavelength is introduced which is then used as a directory name to search for the input files in. It is also attached to the final output nc file. the variable depends on the settings given to libradtran_write_input_file.py

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>