Update computation of pressure on model full and half levels
The calculation of the full level pressure was using the same approach as for the calculation of the half levels. It was adjusted according to: https://confluence.ecmwf.int/display/CKB/ERA5%3A+compute+pressure+and+geopotential+on+model+levels%2C+geopotential+height+and+geometric+height
Now the mean between the half level above and below is used as the full level pressure. This causes differences in the lower levels below 0.05 Pa.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>