Add baran2017 and fu ice parameterization as given in ecRad source code
Use importlib_resources to use netcdf files with coefficients for parameterizations.

The coefficients in the fu netCDF file had to be renamed to be able to read the file in with xarray. They were two dimensional but the second dimension was named the same as the variable which causes a conflict.

Add the licensing terms to the functions copied from the ecRad source code.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>