Adjust calculate_pressure_height to take in multidimensional data
- add the calculation of pressure height in ecrad_read_ifs.py
- add type hinting to barometric_height function

Up to now the function had to be given a two dimensional dataset (time and height). Now it can take any dimensionality as long as time and height are also present. Height is represented with the dimensions half_level and level. A check ensures that the dimension is present before attempting to calculate the pressure height.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>