Update search for nearest neighbor
The kdTree calculates the distance between two points on a plane. However, we are on a sphere and thus need to use the Haversine difference (for example). We therefore switch to the BallTree from scikit learn which lets us use the haversine distance.

Additionally, we include the distance between the aircraft and the model grid points in the output file.

We start counting from 0 for the column dimension.

As there appear errors when writing the files in parallel we remove this again.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>