Update sea ice albedo calculation
Instead of using the value closest to the actual date, lineraly interpolate the sea ice albedo to the date. This causes no difference for the simulations in April as the sea ice albedo does not change between March and Mai.

To facilitate an easy interpolation create a DataArray of the sea ice albedo parameterization in helpers.py.

Signed-off-by: Johannes Röttenbacher <johannes.roettenbacher@web.de>